remove Python 3.8 support
Python 3.8 produces a type error if we use OrderedDict with types for key and value. There is no easy solution for the problem except do not use type hints but the hints brings some benefits. Therefor and because the project is quiet new and do not ues Python 3.8 in our CI, I drop the support for Python 3.8.