filter combination Cxx + CUDA backend + host compiler GCC or Clang
If we know that the CUDA backend is enabled and the host compiler is GCC or Clang, the device compiler must be Nvcc. In this case, we know that the CUDA SDK and Nvcc has the same version number and therefore we can determine which is the maximum supported C++ standard.